# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GridTrade renewable-energy intelligence and P2P coordination API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  degraded: 'degraded',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  service: string;
  timestamp: string;
}

export type GridDecision = typeof GridDecision[keyof typeof GridDecision];


export const GridDecision = {
  APPROVED: 'APPROVED',
  ADJUSTED: 'ADJUSTED',
  RESTRICTED: 'RESTRICTED',
} as const;

export interface EnergyObservation {
  timestamp: string;
  generationKwh: number;
  consumptionKwh: number;
  marketableSurplusKwh: number;
}

export interface DashboardSummary {
  marketableSurplusKwh: number;
  activeListings: number;
  activeProsumers: number;
  currentPriceInrPerKwh: number;
  gridDecision: GridDecision;
  gridLabel: string;
  carbonAvoidedKg: number;
  energyTrend: EnergyObservation[];
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  listing: 'listing',
  match: 'match',
  grid: 'grid',
  transaction: 'transaction',
  prediction: 'prediction',
  alert: 'alert',
} as const;

export type ActivityEventStatus = typeof ActivityEventStatus[keyof typeof ActivityEventStatus];


export const ActivityEventStatus = {
  positive: 'positive',
  neutral: 'neutral',
  warning: 'warning',
} as const;

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  title: string;
  detail: string;
  occurredAt: string;
  status: ActivityEventStatus;
}

export type SolarSystemStatus = typeof SolarSystemStatus[keyof typeof SolarSystemStatus];


export const SolarSystemStatus = {
  online: 'online',
  standby: 'standby',
  offline: 'offline',
} as const;

export interface SolarSystem {
  id: string;
  name: string;
  capacityKw: number;
  location: string;
  status: SolarSystemStatus;
  todayGenerationKwh: number;
  todayConsumptionKwh: number;
}

export interface SolarSystemInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @exclusiveMinimum 0 */
  capacityKw: number;
  /**
     * @minLength 1
     * @maxLength 180
     */
  location: string;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  active: 'active',
  matched: 'matched',
  closed: 'closed',
  restricted: 'restricted',
} as const;

export interface Listing {
  id: string;
  sellerName: string;
  location: string;
  quantityKwh: number;
  priceInrPerKwh: number;
  status: ListingStatus;
  availableFrom: string;
  availableUntil: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  matchScore: number;
  gridDecision: GridDecision;
}

export interface ListingInput {
  solarSystemId: string;
  /** @exclusiveMinimum 0 */
  quantityKwh: number;
  /** @exclusiveMinimum 0 */
  priceInrPerKwh: number;
  availableFrom: string;
  availableUntil: string;
}

export interface GridStatus {
  decision: GridDecision;
  label: string;
  congestionPercent: number;
  renewableSharePercent: number;
  frequencyHz: number;
  updatedAt: string;
}

export interface MatchRecommendation {
  id: string;
  listingId: string;
  sellerName: string;
  location: string;
  quantityKwh: number;
  priceInrPerKwh: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  rationale: string;
  factors: string[];
}

export type AuthSessionRole = typeof AuthSessionRole[keyof typeof AuthSessionRole];


export const AuthSessionRole = {
  PROSUMER: 'PROSUMER',
  CONSUMER: 'CONSUMER',
  UTILITY: 'UTILITY',
  REGULATOR: 'REGULATOR',
  ADMIN: 'ADMIN',
} as const;

export interface AuthSession {
  authenticated: boolean;
  role: AuthSessionRole;
  displayName: string;
}

export type LimitParameter = number;

export type WindowParameter = typeof WindowParameter[keyof typeof WindowParameter];


export const WindowParameter = {
  '24h': '24h',
  '7d': '7d',
  '30d': '30d',
} as const;

export type GetActivityFeedParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: LimitParameter;
};

export type GetEnergyOverviewParams = {
window?: WindowParameter;
};

export type ListListingsParams = {
status?: ListingStatus;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};