# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GridTrade renewable-energy intelligence and P2P coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { GridDecision } from './gridDecision';
import type { ListingStatus } from './listingStatus';

export interface Listing {
  id: string;
  sellerName: string;
  location: string;
  quantityKwh: number;
  priceInrPerKwh: number;
  status: ListingStatus;
  availableFrom: Date;
  availableUntil: Date;
  /**
     * @minimum 0
     * @maximum 100
     */
  matchScore: number;
  gridDecision: GridDecision;
}