# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GridTrade renewable-energy intelligence and P2P coordination API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  solarSystemId: string;
  /** @exclusiveMinimum 0 */
  quantityKwh: number;
  /** @exclusiveMinimum 0 */
  priceInrPerKwh: number;
  availableFrom: Date;
  availableUntil: Date;
}