# lib/api-zod/src/generated/types/listingStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GridTrade renewable-energy intelligence and P2P coordination API
 * OpenAPI spec version: 0.1.0
 */

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  active: 'active',
  matched: 'matched',
  closed: 'closed',
  restricted: 'restricted',
} as const;