# lib/api-zod/src/generated/types/matchRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GridTrade renewable-energy intelligence and P2P coordination API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchRecommendation {
  id: string;
  listingId: string;
  sellerName: string;
  location: string;
  quantityKwh: number;
  priceInrPerKwh: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  rationale: string;
  factors: string[];
}